pyproject.toml
[project]
name = "azctx"
dynamic = ["version"]
description = "A personal Python CLI tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "questionary>=2.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",  # Fast Python linter and formatter (optional per constitution)
]

[project.scripts]
azctx = "src.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pyinstaller>=6.16.0",
]
